Fragment Meter into 4 modules
Was getting way too long and way too complex!
